pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "ai-moderation-eval"
version = "0.1.0"
description = "Production-grade evaluation framework for auditing AI content moderation systems on edge-case accuracy and human-AI handoff quality."
readme = "README.md"
requires-python = ">=3.11"
license = "MIT"
authors = [
    { name = "Yingshi Liu" },
]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
]
dependencies = [
    "pyyaml>=6.0",
    "httpx>=0.27",
    "openai>=1.30",
    "anthropic>=0.25",
    "jsonlines>=4.0",
    "jinja2>=3.1",
    "rich>=13.0",
    "click>=8.1",
    "numpy>=1.26",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.23",
    "ruff>=0.4",
    "mypy>=1.10",
]

[project.scripts]
run-eval = "scripts.run_eval:main"

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W", "UP"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
